Move iCloud resync into init
