+++ app/clients/icloud/routes/site/index.js
@@ -5,7 +5,7 @@ const email = require("helper/email");
 const maxFileSize = config.icloud.maxFileSize; // Maximum file size for iCloud uploads in bytes
 const limit = `${maxFileSize / 1000000}mb`; // limit must be in the format '5mb'
 
-const resyncRecentlySynced = require("../../util/resyncRecentlySynced");
+const { resyncRecentlySynced } = require("../../init");
 const site = new express.Router();
 
 site.use(require("./middleware/authorize"));
